refactor(models): move PlanTaskInput/PlanSpecInput to models/plan.py
Extract data classes from plan_agent.py to comply with file size limit (400 lines).

**Changes**:
- Move PlanTaskInput and PlanSpecInput to models/plan.py
- plan_agent.py: 417 → 396 lines (-21 lines)
- models/plan.py: 52 → 71 lines (+19 lines)

**Rationale**:
- Data models belong in models/ directory
- Reduces plan_agent.py to under 400 line limit
- Improves code organization and cohesion
- All tests passing (25/25)
